Refactored the engine manager
